+++ Heuristics/CornerCapture.cpp
@@ -1,6 +1,6 @@
 #include "CornerCapture.h"
 
-int CornerCapture::evaluate(const std::vector<std::vector<int>>& board) {
+double CornerCapture::evaluate(const std::vector<std::vector<int>>& board) {
     int currentPlayer = getPlayer();
     int opponent = currentPlayer * -1;
     int maxCount, minCount;
@@ -11,7 +11,7 @@ int CornerCapture::evaluate(const std::vector<std::vector<int>>& board) {
         if(board[corner[0]][corner[1]] == opponent) minCount++;
     }
 
-    int heuristic_value = 0;
+    double heuristic_value = 0;
     if(maxCount + minCount != 0) {
         heuristic_value = (maxCount - minCount) * 100 / (maxCount + minCount);
     }
